Redux moved to types/api
