+++ frontend/src/metabase-types/types/Table.ts
@@ -2,10 +2,10 @@ import { Field } from "./Field";
 import { Metric } from "./Metric";
 import { DatabaseId } from "./Database";
 
-import { ISO8601Time } from "metabase-types/api/time";
 import { ForeignKey } from "metabase-types/api/foreignKey";
 import { Segment } from "metabase-types/api/segment";
 import { SchemaName, TableId } from "metabase-types/api/table";
+import { ISO8601Time } from "metabase-types/api/time";
 
 type TableVisibilityType = string; // FIXME
